+++ Examples/VideoPlayer/.gitignore
@@ -1,6 +1,3 @@
-VideoPlayer.xcodeproj/project.xcworkspace/**/*
-VideoPlayer.xcodeproj/xcuserdata/**/*
-
 # OSX
 #
 .DS_Store
@@ -27,6 +24,7 @@ project.xcworkspace
 
 # Android/IJ
 #
+*.iml
 .idea
 .gradle
 local.properties
@@ -35,3 +33,9 @@ local.properties
 #
 node_modules/
 npm-debug.log
+
+# BUCK
+buck-out/
+\.buckd/
+android/app/libs
+android/keystores/debug.keystore
