+++ Examples/VideoPlayer/android/app/src/main/res/values/strings.xml
@@ -1,3 +1,4 @@
 <resources>
+
     <string name="app_name">VideoPlayer</string>
 </resources>
