+++ Examples/VideoPlayer/iOS/VideoPlayerTests/VideoPlayerTests.m
@@ -13,7 +13,7 @@
 #import "RCTLog.h"
 #import "RCTRootView.h"
 
-#define TIMEOUT_SECONDS 240
+#define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React Native!"
 
 @interface VideoPlayerTests : XCTestCase
