+++ Examples/VideoPlayer/package.json
@@ -6,8 +6,8 @@
     "start": "node_modules/react-native/packager/packager.sh"
   },
   "dependencies": {
-    "react": "^0.14.8",
-    "react-native": "0.23.0",
+    "react": "15.3.1",
+    "react-native": "^0.33.0",
     "react-native-video": "../../"
   }
 }
