+++ android/src/main/java/com/brentvatne/react/ReactVideoViewManager.java
@@ -82,8 +82,8 @@ public class ReactVideoViewManager extends SimpleViewManager<ReactVideoView> {
             videoView.setSrc(
                     src.getString(PROP_SRC_URI),
                     src.getString(PROP_SRC_TYPE),
-                    src.getBoolean(PROP_SRC_IS_NETWORK),,
-                    src.getBoolean(PROP_SRC_IS_ASSET)
+                    src.getBoolean(PROP_SRC_IS_NETWORK),
+                    src.getBoolean(PROP_SRC_IS_ASSET),
                     mainVer,
                     patchVer
             );
